VEOSS-293 use currency field alongside using bigdecimal with basic validation

+++ app/pages/NetValueOfSalesFromNiPage.scala
@@ -22,7 +22,7 @@ import PageConstants._
 import play.api.libs.json.JsPath
 import play.api.mvc.Call
 
-case class NetValueOfSalesFromNiPage(countryIndex: Index, vatRateIndex: Index) extends QuestionPage[Int] {
+case class NetValueOfSalesFromNiPage(countryIndex: Index, vatRateIndex: Index) extends QuestionPage[BigDecimal] {
 
   override def path: JsPath = JsPath \ salesFromNi \ countryIndex.position \ salesAtVatRate \ vatRateIndex.position \ toString
